Committed 2021/06/04

+++ Python Machine Learning by Example/Chapter5 - Logistic Regression/Implementing logistic regression using TensorFlow_10.py
@@ -70,7 +70,7 @@ for step, (batch_x, batch_y) in enumerate(train_data.take(training_steps), 1):
         loss = tf.reduce_mean(tf.nn.sigmoid_cross_entropy_with_logits(labels=batch_y, logits=logits))
         print("step: %i, loss: %f" % (step, loss))
 
-X_test_enc = np.array(X_test_enc, dtype=np.float32)
+X_test_enc = np.array(X_test_enc, dtype=np.int32)
 
 logits = tf.add(tf.matmul(X_test_enc.shape[0], weight), bias)[:, 0]
 prediction = tf.nn.sigmoid(logits)
